### 2023-04-03 e5ae85b
add missing DOIs
