+++ paper.bib
@@ -5,7 +5,8 @@
   volume={47},
   pages={1--12},
   year={2015},
-  publisher={Springer}
+  publisher={Springer},
+  DOI={10.3758/s13428-014-0458-y}
 }
 
 @misc{deleeuw2023datapipe,
@@ -25,7 +26,8 @@
   volume={44},
   pages={314--324},
   year={2012},
-  publisher={Springer}
+  publisher={Springer},
+  DOI={10.3758/s13428-011-0168-7}
 }
 
 
@@ -54,7 +56,8 @@
   volume={17},
   number={3},
   pages={299--311},
-  year={2021}
+  year={2021},
+  DOI={10.20982/tqmp.17.3.p299}
 }
 
 @article{strittmatter2022random,
@@ -64,7 +67,8 @@
   pages={883--898},
   volume={55},
   year={2022},
-  publisher={Springer}
+  publisher={Springer},
+  DOI={10.3758/s13428-021-01767-3}
 }
 
 
@@ -76,7 +80,8 @@
   pages={508--515 },
   volume={55},
   year={2023},
-  publisher={Springer}
+  publisher={Springer},
+  DOI={10.3758/s13428-022-01803-w}
 }
 
 
@@ -87,7 +92,8 @@
   volume={53},
   pages={301--310},
   year={2021},
-  publisher={Springer}
+  publisher={Springer},
+  DOI={10.3758/s13428-020-01445-w}
 }
 
 @article{rajananda2018random,
@@ -95,7 +101,8 @@
   author={Rajananda, Sivananda and Lau, HW and Odegaard, Brian},
   journal={Journal of Open Research Software},
   year={2018},
-  publisher={Ubiquity Press Ltd.}
+  publisher={Ubiquity Press Ltd.},
+  DOI={10.1101/192377}
 }
 
 @article{yang2021webcam,
@@ -106,7 +113,8 @@
   number={6},
   pages={1485--1505},
   year={2021},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  DOI={10.1017/S1930297500008512}
 }
 
 @article{hartshorne2019thousand,
@@ -116,7 +124,8 @@
   volume={51},
   pages={1782--1803},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  DOI={10.3758/s13428-018-1155-z}
 }
 
 @article{henninger2021lab,
@@ -126,7 +135,8 @@
   pages={556--573},
   volume={54},
   year={2021},
-  publisher={Springer}
+  publisher={Springer},
+  DOI={10.3758/s13428-019-01283-5}
 }
 
 @article{anwyl2020gorilla,
@@ -136,7 +146,8 @@
   volume={52},
   pages={388--407},
   year={2020},
-  publisher={Springer}
+  publisher={Springer},
+  DOI={10.3758/s13428-019-01237-x}
 }
 
 @article{de2022simulating,
@@ -145,7 +156,8 @@
   journal={Behavior Research Methods},
   pages={1--11},
   year={2022},
-  publisher={Springer}
+  publisher={Springer},
+  DOI={10.3758/s13428-022-01899-0}
 }
 
 @article{kuroki2022jsquestplus,
@@ -154,7 +166,8 @@
   journal={Behavior Research Methods},
   pages={1--8},
   year={2022},
-  publisher={Springer}
+  publisher={Springer},
+  DOI={10.3758/s13428-022-01948-8}
 }
 
 @book{peirce2022building,
@@ -183,7 +196,8 @@
   number={1},
   pages={4--14},
   year={2017},
[…]
