+++ paper.bib
@@ -185,7 +185,8 @@
   number={1},
   pages={24--31},
   year={2017},
-  publisher={Sage Publications Sage CA: Los Angeles, CA}
+  publisher={Sage Publications Sage CA: Los Angeles, CA},
+  DOI={10.1177/0098628316677643}
 }
 
 @article{scott2017lookit,
