### 2023-04-03 #5330 @jodeleeuw
I've gone ahead and fixed the missing DOIs identified above.

Thanks @oliviaguest for editing this!
