### 2023-04-08 #5351 @editorialbot
[REVIEW]: jsPsych: Enabling an Open-Source Collaborative Ecosystem of Behavioral Experiments
**Submitting author:** @jodeleeuw (<a href="http://orcid.org/0000-0003-4815-2364">Joshua de Leeuw</a>)
**Repository:** https://github.com/jspsych/jspsych
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v7.3.2
