### 2023-04-08 #5351 @oliviaguest
👋 @jodeleeuw, this is where the review will actually take place. @pmcharrison, @xinyiguan, @chartgerink please use this issue to leave your comments and feedback for the authors (and please read all the instructions above) — however:

* If you need to comment on [the code itself](https://github.com/jspsych/jspsych) directly (say like making a PR or opening an issue), please link to the PR/issue you create here, so I can get to it from this issue.
* If you need to give feedback on the paper directly (also feel free to make a PR if need be, like for typos, say), leave the comments here (or link to them), so I can see them from here easily.
* If you have to comment on broad overarching things, like any feedback generally, just leave the comments here.

Hope this is clear, let me know if not, and thank you for your time! ☺️
