### 2023-04-12 #5351 @pmcharrison
@jodeleeuw looking good! great that jsPsych version 7 will be credited in this way!

Regarding 'state of the field', I wonder if you can add a touch more here. This is the main comparison with previous work:

> jsPsych and these other options vary in ways such as available features, closed vs. open source, primary programming language, and syntax/style choices, but the main distinction is the particular way that jsPsych abstracts the design of an experiment. jsPsych experiments are constructed using plugins — self-contained modules that define an event and its parameters.

This is a positive statement about what _jsPsych_ does in terms of plugins, but maybe you can spare a few words to say whether/which _other_ programs have analogous constructs to plugins?

Regarding the references, I just noticed that some of the capitalisation hasn't come through (e.g. r package, pushkin).
