+++ paper.bib
@@ -50,7 +50,7 @@
 
 
 @article{gibeau2021corsi,
-  title={The Corsi Blocks Task: Variations and coding with jsPsych},
+  title={The {C}orsi Blocks Task: Variations and coding with jsPsych},
   author={Gibeau, R},
   journal={The Quantitative Methods for Psychology},
   volume={17},
@@ -74,7 +74,7 @@
 
 
 @article{donhauser2022audio,
-  title={Audio-Tokens: A toolbox for rating, sorting and comparing audio samples in the browser},
+  title={Audio-{T}okens: A toolbox for rating, sorting and comparing audio samples in the browser},
   author={Donhauser, Peter W and Klein, Denise},
   journal={Behavior Research Methods},
   pages={508--515 },
@@ -118,7 +118,7 @@
 }
 
 @article{hartshorne2019thousand,
-  title={A thousand studies for the price of one: Accelerating psychological science with Pushkin},
+  title={A thousand studies for the price of one: Accelerating psychological science with {P}ushkin},
   author={Hartshorne, Joshua K and de Leeuw, Joshua R and Goodman, Noah D and Jennings, Mariela and O’Donnell, Timothy J},
   journal={Behavior research methods},
   volume={51},
@@ -129,7 +129,7 @@
 }
 
 @article{henninger2021lab,
-  title={lab. js: A free, open, online study builder},
+  title={lab.js: A free, open, online study builder},
   author={Henninger, Felix and Shevchenko, Yury and Mertens, Ulf K and Kieslich, Pascal J and Hilbig, Benjamin E},
   journal={Behavior Research Methods},
   pages={556--573},
@@ -231,7 +231,7 @@
 }
 
 @article{harrison2020psychtestr,
-  title={psychTestR: An R package for designing and conducting behavioural psychological experiments},
+  title={psychTestR: An {R} package for designing and conducting behavioural psychological experiments},
   author={Harrison, Peter},
   journal={The Journal of Open Source Software},
   volume={5},
