### 2023-04-13 #5351 @jodeleeuw
Thank you @pmcharrison! I've fixed the reference formatting issues.

We definitely want to correctly characterize the state of field.

I think one issue is getting clear about what we think the distinction is between plugins and other ways of abstracting experiment design. `psychTestR` has probably the closest analog to plugins, so it would be helpful to get your take on what is an accurate description of any differences that do exist. Here's my current thinking:

- Some software (lab.js, psychopy, opensesame, gorilla, nodegame, penncontroller, psytoolkit) primarily abstracts the design at a different level, providing a set of resuable components that can be deployed in a different tasks, like image displays, buttons, keyboard control, etc.
- Some software (empirica) borrows abstraction from more general-purpose tools like React to create resuable components that could exist at different levels. I could create a component for a button, or a component for a whole task. Components can nest.
- Some software (psychtestr, lookit) abstracts the design in a very similar way to jsPsych, focusing on parameterized events that are arranged in some sequence. jsPsych is slightly different from these options in two ways (and this could of course change quickly!): (1) jsPsych plugins are fully modular, existing as separate files. This is relevant because it facilitates distributed development, where one plugin can be updated independent of a new version of the core library. (2) There is infrastructure in place, especially with version 7, to facilitate distributed development and sharing. These are not significant differences in terms of underlying design, since I think it wouldn't be difficult to create a new task type for `psychTestR` and publish it as it's own package, and I think the same is true of Lookit. So I don't want to claim that this is a big distinction, but rather just a difference in how things are currently setup to facilitate the kind of collaborative ecosystem that we are emphasizing in this paper.

Does that seem reasonable to you? If so, I'll work on a more concise description for the paper. If not, I'm happy to get feedback!
