### 2023-04-14 #5351 @pmcharrison
Hi @jodeleeuw, many thanks for this interesting overview! It's very helpful to see it laid out in this way.

From you say, I think I agree that the best thing is to emphasise the 'community library' aspect of this. The software mentioned in the first category does contain analogous constructions to plugins under the hood, but it's not so easy to contribute your own unless you're a core developer. You're right that psychTestR in theory allows such contributions, but in practice the vast majority of psychTestR community contributions correspond to entire tests (e.g. a particular IQ test, a particular questionnaire) rather than new response interfaces. In contrast, the jsPsych community library provides an unparalleled source of interfaces.

Maybe all that's needed is something like 'While many psychology experiment frameworks contain abstractions analogous to plugins, it is typically hard or impossible for users to contribute their own plugins. In contrast, we have worked hard to develop a system for jsPsych that makes it easy for users to develop their own plugins and share them with other psychologists via open-source repositories. Our community library already contains X plugins, ranging from Y to Z, which ....' What do you think?
