### 2023-04-21 0f0f158
clarify how plugins do/do not differ from existing solutions
