### 2023-04-21 #5351 @jodeleeuw
@pmcharrison thank you for the very helpful suggestion. We borrowed chunks of it and reworked the statement of need section a bit. Here's [the commit](https://github.com/jspsych/jsPsych/commit/0f0f1589992f68689262d02b9c79670d8815092d) so you can see exactly what changed, though I guess it is all one big line in markdown so the actual changes aren't nicely highlighted.
