### 2023-04-22 #5351 @xinyiguan
@jodeleeuw Thank you for this very cool development! I think jsPsych is a valuable contribution to the community, and I particularly appreciate the community-driven and self-contained modular design aspects of the software. The latest version, incorporating feedback from @pmcharrison, looks excellent and has already addressed the points I wanted to raise in my previous review draft.

@oliviaguest I have completed my checklist!
