### 2023-05-08 #5351 @jodeleeuw
Yep! It's already archived on Zenodo. Here's the DOI for v7.3.2: https://doi.org/10.5281/zenodo.7702307

And, yes, v7.3.2 is the current release.

Thank you!
