+++ docs/about/about.md
@@ -6,6 +6,12 @@ jsPsych was created by [Josh de Leeuw :fontawesome-brands-twitter:](http://www.t
 
 ### Citation
 
-If you use jsPsych please cite the following paper.
+Citations help us demonstrate that jsPsych is used and valued, which allows us to continue working on it. If you use this library in academic work, please cite one of the papers that describes jsPsych.
 
-de Leeuw, J. R. (2015). jsPsych: A JavaScript library for creating behavioral experiments in a web browser. _Behavior Research Methods_, _47_(1), 1-12. [doi:10.3758/s13428-014-0458-y](https://doi.org/10.3758/s13428-014-0458-y).
+For jsPsych versions 6 and earlier, the preferred citation is:
+
+de Leeuw, J.R. (2015). jsPsych: A JavaScript library for creating behavioral experiments in a Web browser. *Behavior Research Methods*, _47_(1), 1-12. doi:[10.3758/s13428-014-0458-y](http://link.springer.com/article/10.3758%2Fs13428-014-0458-y)
+
+For jsPsych version 7, the preferred citation is:
+
+de Leeuw, J.R., Gilbert, R.A., & Luchterhandt, B. (2023). jsPsych: Enabling an open-source collaborative ecosystem of behavioral experiments. *Journal of Open Source Software*, 
